+++ src/main/java/mod/tjt01/lapislib/block/OptionalBlock.java
@@ -4,6 +4,7 @@ import net.minecraft.ChatFormatting;
 import net.minecraft.core.NonNullList;
 import net.minecraft.network.chat.Component;
 
+import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
@@ -11,6 +12,7 @@ import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.Block;
 
 import javax.annotation.Nullable;
+import java.util.HashMap;
 import java.util.List;
 import java.util.function.Supplier;
 
@@ -28,6 +30,10 @@ public class OptionalBlock extends Block {
         return "lapislib.common.disabled";
     }
 
+    public Component getDisabledHoverText() {
+        return Component.translatable(getDisabledTooltip()).withStyle(ChatFormatting.RED);
+    }
+
     @Override
     public void fillItemCategory(CreativeModeTab group, NonNullList<ItemStack> stacks) {
         if (condition.get())
@@ -37,7 +43,7 @@ public class OptionalBlock extends Block {
     @Override
     public void appendHoverText(ItemStack stack, @Nullable BlockGetter getter, List<Component> components, TooltipFlag flag) {
         if (!condition.get())
-            components.add(Component.translatable(this.getDisabledTooltip()).withStyle(ChatFormatting.RED));
+            components.add(getDisabledHoverText());
         super.appendHoverText(stack, getter, components, flag);
     }
 }