+++ src/main/java/mod/tjt01/lapislib/core/datagen/package-info.java
@@ -1,4 +1,12 @@
 @ApiStatus.Internal
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+@FieldsAreNonnullByDefault
+
 package mod.tjt01.lapislib.core.datagen;
 
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import org.jetbrains.annotations.ApiStatus;
+
+import javax.annotation.ParametersAreNonnullByDefault;