+++ src/main/java/mod/tjt01/lapislib/core/network/package-info.java
@@ -1,4 +1,11 @@
 @ApiStatus.Internal
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+@FieldsAreNonnullByDefault
 package mod.tjt01.lapislib.core.network;
 
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import org.jetbrains.annotations.ApiStatus;
+
+import javax.annotation.ParametersAreNonnullByDefault;