+++ src/main/java/mod/tjt01/lapislib/core/package-info.java
@@ -3,6 +3,13 @@
  */
 
 @ApiStatus.Internal
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+@FieldsAreNonnullByDefault
 package mod.tjt01.lapislib.core;
 
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
 import org.jetbrains.annotations.ApiStatus;
+
+import javax.annotation.ParametersAreNonnullByDefault;