+++ src/main/java/mod/tjt01/lapislib/util/client/package-info.java
@@ -1,11 +1,9 @@
-@ApiStatus.Internal
 @MethodsReturnNonnullByDefault
 @ParametersAreNonnullByDefault
 @FieldsAreNonnullByDefault
-package mod.tjt01.lapislib.core.event;
+package mod.tjt01.lapislib.util.client;
 
 import net.minecraft.FieldsAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
-import org.jetbrains.annotations.ApiStatus;
 
 import javax.annotation.ParametersAreNonnullByDefault;