+++ src/test/java/mod/tjt01/lapislibtest/item/LapisLibTestItems.java
@@ -2,6 +2,7 @@ package mod.tjt01.lapislibtest.item;
 
 import mod.tjt01.lapislib.item.OptionalItem;
 import mod.tjt01.lapislibtest.LapisLibTest;
+import mod.tjt01.lapislibtest.config.LapisLibTestConfig;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.Item;
 import net.minecraftforge.eventbus.api.IEventBus;
@@ -12,7 +13,7 @@ import net.minecraftforge.registries.RegistryObject;
 public class LapisLibTestItems {
     public static final DeferredRegister<Item> ITEMS = DeferredRegister.create(ForgeRegistries.ITEMS, LapisLibTest.MODID);
 
-    public static final RegistryObject<OptionalItem> OPTIONAL_ITEM = ITEMS.register("optional_item", () -> new OptionalItem(new Item.Properties().tab(CreativeModeTab.TAB_MISC), () -> false));
+    public static final RegistryObject<OptionalItem> OPTIONAL_ITEM = ITEMS.register("optional_item", () -> new OptionalItem(new Item.Properties().tab(CreativeModeTab.TAB_MISC), () -> LapisLibTestConfig.enableOptionalItems));
 
     public static void register(IEventBus bus){
         ITEMS.register(bus);