+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The version format is MCMAJOR.MCMINOR.MCPATCH-MAJOR.MINOR.PATCH, eg, 1.18.1-1.2.
 ## [UNRELEASED]
 ### Added
 - Ported to 1.19.2 🐸
+- Added `OptionalItem#getDisabledHoverText` and `OptionalBlock#getDisabledHoverText`, which return Components instead of Strings
 - Added global loot modifier types `lapislib:replace_item`, `lapislib:remove_item`, and `lapislib:add_entry`
 - Added loot condition `lapislib:block_tag`
 - Added configuration gui api